コンポーネントの切り分け
