+++ components/Calender/CalendarComponentDaysParts.tsx
@@ -1,0 +1,84 @@
+import React from 'react';
+import {createCalendarData} from '../../utils/utils'; 
+import AddFormComponent from '../Modal/AddForm';
+
+interface Props {
+  year: number;
+  month: number;
+  selectedWeek : number
+  selectedOption : "month" | "week"
+  isAdd : boolean
+  setIsAdd : React.Dispatch<React.SetStateAction<boolean>>;
+}
+
+const CalendarComponent = ({ year, month, selectedWeek ,selectedOption, isAdd,setIsAdd } : Props) => {
+  // カレンダーのデータを生成
+  const { day, end_of_month } = createCalendarData(year, month);
+  const cal : number[][] = make_cal(day, end_of_month);
+
+  const today = new Date();
+  const todayDate = today.getDate();
+  const currentMonth = today.getMonth() + 1;
+  const currentYear = today.getFullYear();
+
+  function make_cal(day: Date, end_of_month: Date): number[][] {
+
+    // 必要な行数（週の数）を計算
+    let firstDayOfWeek : number = new Date(day.getFullYear(), day.getMonth(), 1).getDay();
+    let weeksInMonth : number = Math.ceil((end_of_month.getDate() + firstDayOfWeek) / 7);
+  
+    // カレンダーの初期化
+    let cal: number[][] = new Array(weeksInMonth).fill(null).map(() => new Array(7).fill(null));
+
+    // 第何週かを管理する
+    let week_num : number = 0; 
+  
+    day.setDate(1);  // 日付を月の初日にリセット
+  
+    for (let i = 1; i <= end_of_month.getDate(); i++) {
+      cal[week_num][day.getDay()] = i;
+      if (day.getDay() === 6) week_num++;
+      day.setDate(day.getDate() + 1);
+    }
+    return cal;
+  }
+
+  return(
+    <>
+      {selectedOption !== "week" ? 
+        (
+          cal.map((week, i) =>
+            week.map((day, j) => {
+              // 今日の日付かどうかを確認
+              const isToday = day === todayDate && month === currentMonth && year === currentYear;
+              return (
+                <div key={`day-${i}-${j}`} className={`w-full h-28 border py-2 border-gray-300 ${isToday ? 'bg-gray-200' : ''} ${i === 0 ? 'border-t-0' : ''}`}>
+                  {day || ''}
+                </div>
+              );
+            })
+          )
+        ) 
+        : 
+        (
+          cal[selectedWeek].map((day, j) => {
+            // 今日の日付かどうかを確認
+            const isToday = day === todayDate && month === currentMonth && year === currentYear;
+            return (
+              <div key={`day-0-${j}`} className={`w-full border-t-0 h-svh border py-2 border-gray-300 ${isToday ? 'bg-gray-200' : ''}`}>
+                {day || ''}
+              </div>
+            );
+          })
+        )
+      }
+      {/* モーダル */}
+      <AddFormComponent
+        isAdd={isAdd}
+        setIsAdd={setIsAdd}
+      />
+    </>
+  )
+}
+
+export default CalendarComponent;
